# WIP: Speed up isel in most cases