@crusaderky 2019-11-14
@shoyer I struggle to understand the meaning of ``Dataset.indexes``. How does it  differ from
```python
{k: v.to_index() for k, v in self._variables.items() if isinstance(v, IndexVariable)
```
?