@shoyer 2019-11-14
Dataset.indexes is currently equivalent to just calling to_index() on each appropriate variable. But eventually we want to separate the notion of an index from variables with names matching dimensions, and this is a prerequisite for that.