@crusaderky 2019-11-15
I would *very* much appreciate if everybody could play around with this PR in their projects and confirm that nothing breaks!