@crusaderky 2019-12-04
Implemented @shoyer 's suggestions and aligned to master.
Merging into master tomorrow!